reset password functionality

+++ Big projects/devcamper_api/controllers/courses.js
@@ -48,11 +48,14 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 // @access  Private
 exports.addCourse = asyncHandler ( async (req, res, next) => {
     req.body.bootcamp = req.params.bootcampId;
+    req.body.user = req.user.id;
     const bootcamp = await Bootcamp.findById(req.params.bootcampId)
 
     if(!bootcamp)
         return next(new ErrorResponse(`No bootcamp with the id of ${req.params.bootcampId}`), 404);
-    
+// Make sure that the bootcamp owner is the user that is logged in
+    if(bootcamp.user.toString !== req.user.id && req.user.role !== 'admin')
+        return next(new ErrorResponse(`User ${req.user.id} is not authorised to add a course to ${bootcamp._id}`, 401));
     const course = await Course.create(req.body);
 
     res.status(200).json({
@@ -65,12 +68,18 @@ exports.addCourse = asyncHandler ( async (req, res, next) => {
 // @route   PUT /api/v1/courses/:id
 // @access  Private
 exports.updateCourse = asyncHandler( async (req, res, next) => {
-    const course = await Course.findByIdAndUpdate(req.params.id, req.body, {
+    let course = await Course.findById(req.params.id)
+
+    if (!course)
+        return next( new ErrorResponse(`No course with the id of ${req.params.id}`), 404)
+    
+// Make sure user is the course owner
+    if(course.user.toString() !== req.user.id && req.user.role !== 'admin')
+        return next(new ErrorResponse(`User ${req.user.id} is not authorised to access the course ${course._id}`, 401));
+    course = await Course.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
         runValidators: true})
     
-    if (!course)
-        return next( new ErrorResponse(`No course with the id of ${req.params.id}`), 404)
     
     res.status(200).json({
         success: true,
@@ -86,6 +95,11 @@ exports.deleteCourse = asyncHandler( async (req, res, next) => {
 
     if(!course) 
         return next(new ErrorResponse(`No course with the id of ${req.params.id}`), 404)
+
+// Make sure user is the course owner
+    if(course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next( new ErrorResponse(`User ${req.user.id} is not authorised to delete course ${course._id}`));
+    }
     
     course.deleteOne();
